+++ .eslintrc.js
@@ -282,7 +282,6 @@ module.exports = {
 						'RangeControl',
 						'SearchControl',
 						'SelectControl',
-						'TextareaControl',
 						'ToggleControl',
 						'ToggleGroupControl',
 						'TreeSelect',