+++ packages/components/src/textarea-control/stories/index.story.tsx
@@ -52,7 +52,6 @@ const Template: StoryFn< typeof TextareaControl > = ( {
 
 export const Default: StoryFn< typeof TextareaControl > = Template.bind( {} );
 Default.args = {
-	__nextHasNoMarginBottom: true,
 	label: 'Text',
 	help: 'Enter some text',
 	placeholder: 'Placeholder',